requirements.txt
functions-framework==3.5.0
cloudevents==1.11.0
google-cloud-logging==3.9.0
google-cloud-aiplatform==1.44.0
google-auth==2.28.2
